Ensure find next/previous buttons don't remain disabled after starting typing
Unfortunately the original commit that introduced this code doesn't really
clarify the reasoning behind that change, but the result is that the find button
state is only updated on opening of the findbar, so if the buttons start
disabled and you start typing, they incorrectly remain disabled unless you close
and re-open the findbar.

At least at a quick glance re-enabling the call to _enableFindButtons() doesn't
seem to have any unwanted side effects when testing in Waterfox Classic.

The resulting behaviour still isn't entirely correct, as I suspect that the
buttons should also remain disabled in the case of "no search result", but it's
still definitively an improvement over the previous behaviour.